documentation and refactoring
